# openapi.yaml
openapi: 3.0.3
info:
  title: image-api
  version: 1.0.0
paths:
  /jobs/image-crop:
    post:
      summary: Create an image-crop job
      operationId: postJobsImageCrop
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageCropRequest'
      responses:
        '200':
          description: Job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /jobs/{id}:
    get:
      summary: Get job status
      operationId: getJobsId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ImageCropRequest:
      type: object
      required:
        - imageUrl
        - x
        - y
        - width
        - height
      properties:
        imageUrl:
          type: string
          format: uri
          minLength: 1
        x:
          type: integer
          minimum: 0
        y:
          type: integer
          minimum: 0
        width:
          type: integer
          minimum: 0
        height:
          type: integer
          minimum: 0
    JobResponse:
      type: object
      required:
        - id
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        croppedImageUrl:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        created_at:
          type: string
        updated_at:
          type: string
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string